+++ GoDiveMVP/Views/Pages/field_guide_marine_life_detail_view.swift
@@ -238,6 +238,7 @@ struct FieldGuideMarineLifeDetailView: View {
                     mediaSightings: taggedSightings,
                     marineLifeCatalog: catalogMarineLifeForPager,
                     ownerProfileID: accountSession.currentProfile?.id,
+                    unitSystem: diveDisplayUnitSystem,
                     bottomScrollInset: bottomScrollInset,
                     onOpenDive: onOpenDive,
                     onOpenSpecies: openCatalogMarineLifeDetail